Document how to quit the suggestions pane
Used to be able to quit the suggestion pane with `C-g`, but that was removed in 397e5c8258c97be8e2abb1d34526219fc816ca86. The the `C-g` was intuitive to me, but `C-c C-k` was not -- I had to look at the change log to figure it out -- so I thought it would be helpful to document it.